+++ tools/testing/run_regression_tests.py
@@ -304,7 +304,7 @@ except Exception as e:
     def test_stm32_search(self):
         """Test 3.3: STM32 Peripheral Search"""
         code = '''
-from circuit_synth.stm32_search_helper import handle_stm32_peripheral_query
+from circuit_synth.ai_integration.stm32_search_helper import handle_stm32_peripheral_query
 result = handle_stm32_peripheral_query('find stm32 with 2 spi available on jlcpcb')
 if result:
     print('✅ STM32 search working')
